Lots of fixes
Separated calculate_result and calculate_error, fixed fault in result calculation, where the result length was longer than the neurons on the next layer (which also caused inability to make more neurons on the next layer than on the previous one). Kinda understood generalized delta rule for backpropagation, wrote an algorithm for it. Added some necessary functions to the class of a Neuron