Continuing the genetic algorithms
Identified some problems (not working when there are no levels in between the in and out), deleted some temporary stuff, added a function create_new_network() to calculate the error for the new population by creating all the necessary networks and doing one cycle on them, Remade the create_initial_population function to create the dictionary, instead of having an unnecessary function create_chromosome_error_dict() for that. Added a small necessary function get_index(self) to the class to be able to copy the networks.