Activation functions and biases
Added biases to the neurons, added the formula to update them while learning. Added a list of activation functions and their derivatives from the activation_functions module, changed the create_network, calculations, genetic_algorithm_functions and neuron_class to accomodate that. Instead of choosing the algorithm, made the code work on both of them one after the other and compare the results. For that created a new function in the calculations module, called main_calculation. Might plan to create a combined genetic-backpropagation algorithm, if I get there. Other small updates, like adding *args not to throw unneeded parameters to the functions, reworking the already existing *args to work better, checking what parameters should functions return and removing the unnecessary ones. Reflecting the changes in the text itself.