Not sure where this is going
Made a function that lists all of the activation functions, had to rename them to keep the alphabetical order. Played around with weight updates in backpropagation. Found out that for linear functions all is broken. Sent crossover and mutation propabilities to main. Fixed a problem of network not working with 0 hidden layers. Tried to get biases of in_layer to 0, need to work on it though. Instead of generating random data now have iris as an example (found real data for the crack problem, will work on it in the next few days). Added a recalculation of the results after training is finished to ensure the netwrok is recalculated on real final weights. Seems not to be working that well due to not really fitting the learning curve. Will check that. Made an Excel file to test the results of the neural network.