Fixing mistakes, making things better
Updated the wrong formulas for backpropagation and the copying of the weight to bias problem for genetic algorithm. Made bias range values not compulsory for the class constructor. Had to change skilearn to load dataset for some reason. Sent population_length to main and passed it as a parameter to the function. Reflected changes in the text itself, added a small user instruction. Possibly have to leave out the plan of recalculating result, since error seems to increase on the last point